src/main/java/thespeace/jdbc/service/MemberServiceV4.java
package thespeace.jdbc.service;

import lombok.extern.slf4j.Slf4j;
import org.springframework.transaction.annotation.Transactional;
import thespeace.jdbc.domain.Member;
import thespeace.jdbc.repository.MemberRepository;

/**
 * <h2>예외 누수 문제 해결</h2>
 * <ul>
 *     <li>SQLException 제거</li>
 *     <li>MemberRepository 인터페이스 의존</li>
 *     <li>순수한 자바코드로 이루어진 서비스 완성</li>
 * </ul>
 */
@Slf4j
public class MemberServiceV4 {

    private final MemberRepository memberRepository;

    public MemberServiceV4(MemberRepository memberRepository) {
        this.memberRepository = memberRepository;
    }

    @Transactional
    public void accountTransfer(String fromId, String toId, int money) {
        bizLogic(fromId, toId, money);
    }

    private void bizLogic(String fromId, String toId, int money) {
        Member fromMember = memberRepository.findById(fromId);
        Member toMember = memberRepository.findById(toId);

        memberRepository.update(fromId, fromMember.getMoney() - money);
        validation(toMember);
        memberRepository.update(toId, toMember.getMoney() + money);
    }

    private void validation(Member toMember) {
        if(toMember.getMemberId().equals("ex")) {
            throw new IllegalStateException("이체 중 예외 발생");
        }
    }
}
